[PL-26058]: Moved classes to correct packages
