+++ 120-ng-manager/modules/ngsettings/BUILD.bazel
@@ -6,31 +6,15 @@ java_library(
     name = "module",
     srcs = glob(["src/main/java/**/*.java"]),
     resources = [],
-    visibility = ["//visibility:public"],
+    visibility = ["//120-ng-manager:__subpackages__"],
     deps = [
         "//:lombok",
+        "//120-ng-manager/modules/ngsettings/contracts:module",
         "//955-outbox-sdk:module",
-        "//960-ng-core-beans:module",
         "//960-persistence:module",
         "//970-ng-commons:module",
         "//980-commons:module",
         "//999-annotations:module",
-        "//120-ng-manager/modules/ngsettings/contracts:module",
-        "@maven//:com_fasterxml_jackson_core_jackson_annotations",
-        "@maven//:com_google_code_findbugs_annotations",
-        "@maven//:com_google_guava_guava",
-        "@maven//:com_google_inject_guice",
-        "@maven//:io_dropwizard_dropwizard_lifecycle",
-        "@maven//:io_swagger_swagger_annotations",
-        "@maven//:javax_validation_validation_api",
-        "@maven//:javax_ws_rs_javax_ws_rs_api",
-        "@maven//:net_jodah_failsafe",
-        "@maven//:org_modelmapper_modelmapper",
-        "@maven//:org_mongodb_morphia_morphia",
-        "@maven//:org_slf4j_slf4j_api",
-        "@maven//:org_springframework_data_spring_data_commons",
-        "@maven//:org_springframework_data_spring_data_mongodb",
-        "@maven//:org_springframework_spring_tx",
     ],
 )
