+++ 120-ng-manager/modules/ngsettings/contracts/BUILD.bazel
@@ -13,13 +13,7 @@ java_library(
         "//970-ng-commons:module",
         "//980-commons:module",
         "//999-annotations:module",
-        "@maven//:com_esotericsoftware_kryo",
-        "@maven//:com_fasterxml_jackson_core_jackson_annotations",
-        "@maven//:com_google_code_findbugs_annotations",
-        "@maven//:com_google_guava_guava",
         "@maven//:io_swagger_swagger_annotations",
-        "@maven//:javax_validation_validation_api",
-        "@maven//:org_springframework_data_spring_data_commons",
     ],
 )
