+++ 120-ng-manager/modules/ngsettings/contracts/src/main/java/io/harness/ngsettings/SettingSource.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+ */
+
+package io.harness.ngsettings;
+
+public enum SettingSource {
+  ACCOUNT,
+  ORG,
+  PROJECT,
+  DEFAULT;
+}
