+++ 120-ng-manager/modules/ngsettings/contracts/src/main/java/io/harness/ngsettings/beans/SettingDTO.java
@@ -1,16 +1,22 @@
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+ */
+
 package io.harness.ngsettings.beans;
 
 import io.harness.NGCommonEntityConstants;
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.data.validator.EntityIdentifier;
-import io.harness.ng.core.setting.SettingCategory;
-import io.harness.ng.core.setting.SettingValueType;
+import io.harness.ngsettings.SettingCategory;
+import io.harness.ngsettings.SettingValueType;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.Set;
 import javax.validation.constraints.NotNull;
-
 import lombok.Builder;
 import lombok.Data;
 import org.hibernate.validator.constraints.NotBlank;
