+++ 120-ng-manager/modules/ngsettings/contracts/src/main/java/io/harness/ngsettings/remote/SettingsResource.java
@@ -1,3 +1,10 @@
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+ */
+
 package io.harness.ngsettings.remote;
 
 import static io.harness.NGCommonEntityConstants.*;
@@ -8,7 +15,7 @@ import io.harness.annotations.dev.OwnedBy;
 import io.harness.ng.core.dto.ErrorDTO;
 import io.harness.ng.core.dto.FailureDTO;
 import io.harness.ng.core.dto.ResponseDTO;
-import io.harness.ng.core.setting.SettingCategory;
+import io.harness.ngsettings.SettingCategory;
 import io.harness.ngsettings.beans.*;
 
 import io.swagger.annotations.Api;
