+++ 120-ng-manager/modules/ngsettings/src/main/java/io/harness/ngsettings/remote/SettingsResourceImpl.java
@@ -1,12 +1,19 @@
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+ */
+
 package io.harness.ngsettings.remote;
 
 import io.harness.ng.core.dto.ResponseDTO;
-import io.harness.ng.core.setting.SettingCategory;
-import io.harness.ngsettings.api.SettingsService;
+import io.harness.ngsettings.SettingCategory;
 import io.harness.ngsettings.beans.SettingRequestDTO;
 import io.harness.ngsettings.beans.SettingResponseDTO;
 import io.harness.ngsettings.beans.SettingValueRequestDTO;
 import io.harness.ngsettings.beans.SettingValueResponseDTO;
+import io.harness.ngsettings.services.SettingsService;
 
 import com.google.inject.Inject;
 import java.util.List;
