+++ 120-ng-manager/modules/ngsettings/src/main/java/io/harness/ngsettings/services/SettingsService.java
@@ -1,6 +1,13 @@
-package io.harness.ngsettings.api;
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+ */
 
-import io.harness.ng.core.setting.SettingCategory;
+package io.harness.ngsettings.services;
+
+import io.harness.ngsettings.SettingCategory;
 import io.harness.ngsettings.beans.SettingRequestDTO;
 import io.harness.ngsettings.beans.SettingResponseDTO;
 import io.harness.ngsettings.beans.SettingValueRequestDTO;
