+++ 120-ng-manager/modules/ngsettings/src/main/java/io/harness/ngsettings/services/impl/SettingsServiceImpl.java
@@ -1,4 +1,11 @@
-package io.harness.ngsettings.api.impl;
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+ */
+
+package io.harness.ngsettings.services.impl;
 
 import static io.harness.annotations.dev.HarnessTeam.PL;
 import static io.harness.outbox.TransactionOutboxModule.OUTBOX_TRANSACTION_TEMPLATE;
@@ -6,21 +13,31 @@ import static io.harness.springdata.TransactionUtils.DEFAULT_TRANSACTION_RETRY_P
 
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.encryption.Scope;
-import io.harness.ng.core.setting.SettingCategory;
-import io.harness.ng.core.setting.SettingSource;
-import io.harness.ng.core.setting.SettingUpdateType;
-import io.harness.ngsettings.api.SettingsService;
-import io.harness.ngsettings.beans.*;
+import io.harness.ngsettings.SettingCategory;
+import io.harness.ngsettings.SettingSource;
+import io.harness.ngsettings.SettingUpdateType;
+import io.harness.ngsettings.beans.SettingDTO;
+import io.harness.ngsettings.beans.SettingRequestDTO;
+import io.harness.ngsettings.beans.SettingResponseDTO;
+import io.harness.ngsettings.beans.SettingValueRequestDTO;
+import io.harness.ngsettings.beans.SettingValueResponseDTO;
 import io.harness.ngsettings.entities.Setting;
 import io.harness.ngsettings.entities.SettingConfiguration;
 import io.harness.ngsettings.mapper.SettingsMapper;
-import io.harness.ngsettings.repositories.SettingConfigurationRepository;
-import io.harness.ngsettings.repositories.SettingsRepository;
+import io.harness.ngsettings.services.SettingsService;
 import io.harness.outbox.api.OutboxService;
+import io.harness.repositories.SettingConfigurationRepository;
+import io.harness.repositories.SettingsRepository;
 
 import com.google.inject.Inject;
 import com.google.inject.name.Named;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
+import java.util.ListIterator;
+import java.util.Map;
+import java.util.Optional;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
@@ -53,9 +70,8 @@ public class SettingsServiceImpl implements SettingsService {
     List<SettingConfiguration> defaultSettings = settingConfigurationRepository.findAllByCategoryAndAllowedScopes(
         category, getScope(orgIdentifier, projectIdentifier));
 
-    List<Setting> settings =
-        settingsRepository.findByAccountIdentifierAndOrgIdentifierAndProjectIdentifierAndCategory(
-            accountIdentifier, orgIdentifier, projectIdentifier, category);
+    List<Setting> settings = settingsRepository.findByAccountIdentifierAndOrgIdentifierAndProjectIdentifierAndCategory(
+        accountIdentifier, orgIdentifier, projectIdentifier, category);
     Map<String, Setting> settingsMap = settings.stream().collect(
         Collectors.toMap(Setting::getIdentifier, Function.identity(), (o, n) -> o, HashMap::new));
 
@@ -166,7 +182,7 @@ public class SettingsServiceImpl implements SettingsService {
   public List<SettingConfiguration> listDefaultSettings() {
     List<SettingConfiguration> settingConfigurationList = new ArrayList<>();
     Iterator<SettingConfiguration> settingConfigurationIterator = settingConfigurationRepository.findAll().iterator();
-    while(settingConfigurationIterator.hasNext()) {
+    while (settingConfigurationIterator.hasNext()) {
       settingConfigurationList.add(settingConfigurationIterator.next());
     }
     return settingConfigurationList;
