+++ 120-ng-manager/modules/ngsettings/src/main/java/io/harness/repositories/SettingsRepository.java
@@ -1,10 +1,17 @@
-package io.harness.ngsettings.repositories;
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+ */
+
+package io.harness.repositories;
 
 import static io.harness.annotations.dev.HarnessTeam.PL;
 
 import io.harness.annotation.HarnessRepo;
 import io.harness.annotations.dev.OwnedBy;
-import io.harness.ng.core.setting.SettingCategory;
+import io.harness.ngsettings.SettingCategory;
 import io.harness.ngsettings.entities.Setting;
 
 import java.util.List;
