+++ 120-ng-manager/src/main/java/io/harness/ng/NextGenModule.java
@@ -216,8 +216,8 @@ import io.harness.ng.webhook.services.api.WebhookEventService;
 import io.harness.ng.webhook.services.api.WebhookService;
 import io.harness.ng.webhook.services.impl.WebhookEventProcessingServiceImpl;
 import io.harness.ng.webhook.services.impl.WebhookServiceImpl;
-import io.harness.ngsettings.api.SettingsService;
-import io.harness.ngsettings.api.impl.SettingsServiceImpl;
+import io.harness.ngsettings.services.SettingsService;
+import io.harness.ngsettings.services.impl.SettingsServiceImpl;
 import io.harness.notification.module.NotificationClientModule;
 import io.harness.opaclient.OpaClientModule;
 import io.harness.outbox.TransactionOutboxModule;
