--- 970-ng-commons/src/main/java/io/harness/ng/core/setting/SettingCategory.java
@@ -1,13 +1,0 @@
-package io.harness.ng.core.setting;
-
-import static io.harness.annotations.dev.HarnessTeam.PL;
-
-import io.harness.annotations.dev.OwnedBy;
-
-@OwnedBy(PL)
-public enum SettingCategory {
-  CD,
-  CI,
-  CCM,
-  CV;
-}
