--- 970-ng-commons/src/main/java/io/harness/ng/core/setting/SettingSource.java
@@ -1,8 +1,0 @@
-package io.harness.ng.core.setting;
-
-public enum SettingSource {
-  ACCOUNT,
-  ORG,
-  PROJECT,
-  DEFAULT;
-}
