--- 970-ng-commons/src/main/java/io/harness/ng/core/setting/SettingUpdateType.java
@@ -1,3 +1,0 @@
-package io.harness.ng.core.setting;
-
-public enum SettingUpdateType { UPDATE, RESTORE }
